Made the program around 10kb smaller and cleaned up the code
