+++ AGG-Productions.csproj
@@ -80,10 +80,8 @@
     </ApplicationDefinition>
     <Compile Include="LauncherData\Json.cs" />
     <Compile Include="LauncherFunctions\Arguments.cs" />
-    <Compile Include="LauncherFunctions\Cache.cs" />
     <Compile Include="LauncherFunctions\Dynamicbuttons.cs" />
-    <Compile Include="LauncherFunctions\GameInstall.cs" />
-    <Compile Include="LauncherFunctions\GameReinstall.cs" />
+    <Compile Include="LauncherFunctions\GameInstallation.cs" />
     <Compile Include="LauncherFunctions\OnlineFunctions.cs" />
     <Compile Include="LauncherFunctions\SelectScreen.cs" />
     <Compile Include="LauncherData\FileDownloader.cs" />
