+++ LauncherFunctions/Dynamicbuttons.cs
@@ -1,18 +1,14 @@
-﻿using AGG_Productions.LauncherData;
+﻿using System;
+using System.IO;
+using System.Net;
+using System.Windows;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Net;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Media;
-using System.Windows.Media.Imaging;
 using static System.Environment;
+using System.Windows.Media.Imaging;
+using AGG_Productions.LauncherData;
 
 namespace AGG_Productions.LauncherFunctions
 {
@@ -20,7 +16,6 @@ namespace AGG_Productions.LauncherFunctions
     {
         public static void SetupButtons()
         {
-            #region Dynamic Buttons
             if (CheckInternet.IsOnline)
             {
                 WebClient d = new WebClient();
@@ -61,7 +56,6 @@ namespace AGG_Productions.LauncherFunctions
                 MainWindow.ListObject.Items.Add(newBtn);
                 newBtn.Click += new RoutedEventHandler(MainWindow.Game_Click);
             }
-            #endregion
         }
     }
 }
